Move more code into subclass.rs.
